🔥 Move tracking-context into tracking
