+++ back/_doc/erreurs.md
@@ -61,10 +61,10 @@
 | [<b>300002</b>](../apps/bridge-http-proxy-rie/src/exceptions/bridge-http-proxy-variable-missing.exception.ts) | Il manque des variables dans la réponse renvoyée par le broker rabbitmq | Une erreur technique est survenue. Si le problème persiste, veuillez nous contacter. | <b>Classe :</b> [BridgeHttpProxyMissingVariableException](../apps/bridge-http-proxy-rie/src/exceptions/bridge-http-proxy-variable-missing.exception.ts) <br /><b>Code HTTP :</b> 500 <br /><b>error :</b> server_error <br /> <b>error_description :</b> authentication aborted due to a technical error on the authorization server |
 | [<b>300003</b>](../apps/bridge-http-proxy-rie/src/exceptions/bridge-http-proxy-csmr.exception.ts) | Une erreur technique est survenue dans le consumer au moment de la récupération des informations à travers le broker rabbitmq | Une erreur technique est survenue. Si le problème persiste, veuillez nous contacter. | <b>Classe :</b> [BridgeHttpProxyCsmrException](../apps/bridge-http-proxy-rie/src/exceptions/bridge-http-proxy-csmr.exception.ts) <br /><b>Code HTTP :</b> 500 <br /><b>error :</b> server_error <br /> <b>error_description :</b> authentication aborted due to a technical error on the authorization server |
 
-## @fc/tracking-context
+## @fc/tracking
 | Code d'erreur | Description | Message utilisateur | Infos techniques |
 |---|---|---|---|
-| [<b>410001</b>](../libs/tracking-context/src/exceptions/tracking-missing-network-context.exception.ts) | L&#39;application n&#39;a pas trouvé de headers dans l&#39;objet request, c&#39;est probablement un bug, Contacter le support N3 | Missing network context (headers) | <b>Classe :</b> [TrackingMissingNetworkContextException](../libs/tracking-context/src/exceptions/tracking-missing-network-context.exception.ts) <br /><b>Code HTTP :</b> 400 <br /><b>error :</b> server_error <br /> <b>error_description :</b> authentication aborted due to a technical error on the authorization server |
+| [<b>410001</b>](../libs/tracking/src/exceptions/tracking-missing-network-context.exception.ts) | L&#39;application n&#39;a pas trouvé de headers dans l&#39;objet request, c&#39;est probablement un bug, Contacter le support N3 | Missing network context (headers) | <b>Classe :</b> [TrackingMissingNetworkContextException](../libs/tracking/src/exceptions/tracking-missing-network-context.exception.ts) <br /><b>Code HTTP :</b> 400 <br /><b>error :</b> server_error <br /> <b>error_description :</b> authentication aborted due to a technical error on the authorization server |
 
 ## @fc/flow-steps
 | Code d'erreur | Description | Message utilisateur | Infos techniques |
