--- back/libs/tracking-context/README.md
@@ -1,7 +1,0 @@
-# Tracking context
-
-Provides helpers functions for generics tracking needs.
-
-Currently available helpers :
-
-- [extractNetworkInfoFromHeaders](./src/helpers/extract-network-context.helper.ts) : Get network info (ip, port, etc.)
