--- back/libs/tracking-context/src/index.ts
@@ -1,2 +1,0 @@
-export * from './helpers';
-export * from './interfaces';
