+++ back/libs/tracking/src/exceptions/tracking-base.exception.ts
