+++ back/libs/tracking/src/index.ts
@@ -1,5 +1,6 @@
 export * from './decorators';
 export * from './dto';
+export * from './helpers';
 export * from './interceptors';
 export * from './interfaces';
 export * from './services';
