+++ back/libs/tracking/src/interfaces/index.ts
@@ -1,6 +1,7 @@
 export * from './core-tracking-context.interface';
 export * from './core-tracking-log.interface';
 export * from './core-tracking-providers.interface';
+export * from './network-context.interface';
 export * from './tracked-event.interface';
 export * from './tracked-event-context.interface';
 export * from './tracked-event-map.interface';
