+++ back/libs/tracking/src/services/core-tracking.service.ts
@@ -3,8 +3,8 @@ import { Injectable } from '@nestjs/common';
 import { overrideWithSourceIfNotNull } from '@fc/common';
 import { UserSession } from '@fc/core';
 import { SessionService } from '@fc/session';
-import { extractNetworkInfoFromHeaders } from '@fc/tracking-context';
 
+import { extractNetworkInfoFromHeaders } from '../helpers';
 import {
   ICoreTrackingContext,
   ICoreTrackingLog,
