+++ back/tsconfig.json
@@ -68,8 +68,6 @@
       "@mocks/email-validator": ["libs/email-validator/.mocks"],
       "@fc/tracking": ["libs/tracking/src"],
       "@fc/tracking/*": ["libs/tracking/src/*"],
-      "@fc/tracking-context": ["libs/tracking-context/src"],
-      "@fc/tracking-context/*": ["libs/tracking-context/src/*"],
       "@fc/app": ["libs/app/src"],
       "@fc/app/*": ["libs/app/src/*"],
       "@fc/notifications": ["libs/notifications/src"],
